some comparison

restore
----
- 1607 manga
- 210620 chapters
- 4 categories (entries: 1588, 6, 7, 13)
- 2 manga with 1 track record each
----
- database empty:
  - current  : 09:56 (596s)
  - optimized: 02:33 (153s)
  - result: ~4x (~74%) faster

- database contains all backup data:
  - current  : 12:25 (745s)
  - optimized: 02:19 (139s)
  - result: ~5x (~81%) faster
----
restore
----
- 130 manga
- 141266 chapters
- 6 categories (entries: 64, 15, 31, 18, 2, 6)
- 102 manga with 1 track records
- 28 manga with 0 track records
----
- database empty:
  - current  : 01:25 (85s)
  - optimized: 00:17 (17s)
  - result: ~5x (~80%) faster
	
- database contains data:
  - current  : 02:36 (156s)
  - optimized: 00:19 (19s)
  - result: ~8x (~87%) faster
  
- database contains all backup data:
  - current  : 02:29 (149s)
  - optimized: 00:17 (17s)
  - result: ~8x (~88%) faster